+++ src/Executor/Elastica/ElasticaExecutor.php
@@ -15,6 +15,7 @@ namespace FiveLab\Component\Ruler\Executor\Elastica;
 
 use Elastica\Query;
 use FiveLab\Component\Ruler\Executor\ExecutorInterface;
+use FiveLab\Component\Ruler\Node\BinaryNode;
 use FiveLab\Component\Ruler\Node\Node;
 use FiveLab\Component\Ruler\Operator\Operators;
 use FiveLab\Component\Ruler\Query\RawSearchQuery;
@@ -30,6 +31,14 @@ readonly class ElasticaExecutor implements ExecutorInterface
 
     public function execute(object $target, Node $node, array $parameters): void
     {
+        if (!$node instanceof BinaryNode) {
+            // A field, a parameter or a constant alone is not a condition: there is no query clause for it.
+            throw new \LogicException(\sprintf(
+                'The rule must be a condition, %s given.',
+                ElasticaVisitor::describeNode($node)
+            ));
+        }
+
         /** @var array<string, mixed> $query */
         $query = $this->visitor->visit($target, $node, $parameters, $this->operators);