+++ tests/Functional/Elastica/ElasticaRulerTest.php
@@ -17,6 +17,7 @@ use Elastica\Query;
 use FiveLab\Component\Ruler\Query\RawSearchQuery;
 use FiveLab\Component\Ruler\Ruler;
 use FiveLab\Component\Ruler\Target\ElasticaTarget;
+use FiveLab\Component\Ruler\Tests\Functional\Elastica\Fixtures\OrderStatus;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\Attributes\TestWith;
@@ -74,6 +75,53 @@ class ElasticaRulerTest extends TestCase
     }
 
     #[Test]
+    #[TestWith([':amount > amount', 'The left side of the operator ">" must be a field, the parameter ":amount" given.'])]
+    #[TestWith(['100 = id', 'The left side of the operator "=" must be a field, the constant "100" given.'])]
+    #[TestWith(['amount > cost', 'The right side of the operator ">" must be a parameter or a constant, the field "cost" given.'])]
+    #[TestWith(['id = :id and published', 'The operator "and" can combine only conditions, the field "published" given.'])]
+    public function shouldFailIfOperandIsNotFieldAndValue(string $rule, string $expectedMessage): void
+    {
+        $this->expectException(\LogicException::class);
+        $this->expectExceptionMessage($expectedMessage);
+
+        $this->ruler->apply(new Query(), $rule, ['id' => 123, 'amount' => 100]);
+    }
+
+    #[Test]
+    #[TestWith(['published', 'The rule must be a condition, the field "published" given.'])]
+    #[TestWith(['(true)', 'The rule must be a condition, the constant "true" given.'])]
+    #[TestWith([':id', 'The rule must be a condition, the parameter ":id" given.'])]
+    public function shouldFailIfRuleIsNotCondition(string $rule, string $expectedMessage): void
+    {
+        $this->expectException(\LogicException::class);
+        $this->expectExceptionMessage($expectedMessage);
+
+        $this->ruler->apply(new RawSearchQuery(), $rule, ['id' => 123]);
+    }
+
+    #[Test]
+    #[DataProvider('provideValuesForNormalize')]
+    public function shouldNormalizeValue(string $rule, array $params, array $expectedQuery): void
+    {
+        $query = new RawSearchQuery();
+
+        $this->ruler->apply($query, $rule, $params);
+
+        self::assertEquals($expectedQuery, $query->toArray()['query']);
+    }
+
+    #[Test]
+    #[TestWith([new Query()])]
+    #[TestWith([new RawSearchQuery()])]
+    public function shouldFailForObjectValue(object $query): void
+    {
+        $this->expectException(\LogicException::class);
+        $this->expectExceptionMessage('The value of the parameter "id" must be a scalar, a date, a backed enum, a stringable object or an array of them, "stdClass" given.');
+
+        $this->ruler->apply($query, 'id = :id', ['id' => new \stdClass()]);
+    }
+
+    #[Test]
     #[TestWith([new Query()])]
     #[TestWith([new RawSearchQuery()])]
     public function shouldCombineQueriesOnRepeatedApply(object $query): void
@@ -103,6 +151,54 @@ class ElasticaRulerTest extends TestCase
 
         self::assertSame(10, $array['size']);
         self::assertEquals(['range' => ['price' => ['gt' => 5]]], $array['query']);
+    }
+
+    public static function provideValuesForNormalize(): array
+    {
+        $stringableId = new class() implements \Stringable {
+            public function __toString(): string
+            {
+                return '9d46f5ca';
+            }
+        };
+
+        return [
+            'date' => [
+                'created > :created',
+                ['created' => new \DateTimeImmutable('2026-01-02 03:04:05.678', new \DateTimeZone('UTC'))],
+                ['range' => ['created' => ['gt' => '2026-01-02T03:04:05.678+00:00']]],
+            ],
+
+            'backed enum' => [
+                'status = :status',
+                ['status' => OrderStatus::Paid],
+                ['bool' => ['must' => [['term' => ['status' => ['value' => 'paid']]]]]],
+            ],
+
+            'stringable' => [
+                'id = :id',
+                ['id' => $stringableId],
+                ['bool' => ['must' => [['term' => ['id' => ['value' => '9d46f5ca']]]]]],
+            ],
+
+            'list with missed keys' => [
+                'id in (:ids)',
+                ['ids' => \array_filter(['123', '', '124'])],
+                ['bool' => ['must' => [['terms' => ['id' => ['123', '124']]]]]],
+            ],
+
+            'list of objects' => [
+                'status in (:statuses)',
+                ['statuses' => [OrderStatus::Paid, $stringableId]],
+                ['bool' => ['must' => [['terms' => ['status' => ['paid', '9d46f5ca']]]]]],
+            ],
+
+            'terms lookup' => [
+                'followers in (:lookup)',
+                ['lookup' => ['index' => 'users', 'id' => $stringableId, 'path' => 'followers']],
+                ['bool' => ['must' => [['terms' => ['followers' => ['index' => 'users', 'id' => '9d46f5ca', 'path' => 'followers']]]]]],
+            ],
+        ];
     }
 
     public static function provideDataForApply(): array